requirements.txt
annotated-types==0.6.0
asgiref==3.8.1
Django==5.0.4
django-debug-toolbar==4.3.0
django-filter==24.2
djangorestframework==3.15.1
greenlet==3.0.3
gunicorn==22.0.0
Mako==1.3.2
MarkupSafe==2.1.5
packaging==24.0
pillow==10.3.0
psycopg2-binary==2.9.9
pydantic==2.7.0
pydantic_core==2.18.1
python-dotenv==1.0.1
SQLAlchemy==2.0.29
sqlparse==0.4.4
typing_extensions==4.11.0
whitenoise==6.6.0
